Fix wrong variable usage.
Having these variables as instance variables instead of having them local
to their specific functions was unnecessary. Also, it was using
memory space for nothing since the variables life depended on the object
hence they'd only be removed from memory when the object is deleted.

Also, another thing noted is the use of the constructor to initialise
instance variables. This is not a requirement, as long as an instance
variable is defined before it is used it's ok. It could be initialised
in the constructor or in a method.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>