Refactoring Account class
Having deposit_amount and withdraw_amount as instance variables is
unnecessary.  They are better off as variables local to their
respective functions.

Also since the variable lower_limit is common to all classes, having
it as a class variable makes more sense. Such that if you want to
change the limit at run time you just need to change it once and it'll
reflect in all objects of that class

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>