Refactor create_account
Removed the instance variable: lower_limit. This is because this
variable is not really related to the bank class but rather to the
account class. Plus it being here was redundant.

Returning the user_account object was not necessary. The initial
assignment to self.user_account was ok.

Added a simple accound id generator just to make things work as
normal.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>