Refactor confirm_pin
Added account_pin as an instance variable in Account. This is because
a pin number is usually associated with an account hence makes
more sense being an instance variable of Account.

It was not necessary to have account_pin_confirm as an instance
variable of Bank. With the constructor being empty, it was removed.

Renamed confirm_pin to get_pin. This better conveys the functions
functionality.

Fixed the issue where inputing a pin less than 4 digits would cause
the loop to exit. This is due to the break in line 105. That was meant
to be a continue statement.

Also fixed the issue of the confirmation running infinitely if the pins
matched.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>