Bug Fix
The break statement in verify_login helps fix an IndexError that gets
thrown. This is a temporary fix as it only works if the user details
entered are in the records.csv

The attribute error in show_main_menu was cause at no point was
the object variable user_account being assigned when the user was
logging in. It was only being assigned when during account creation.
This commit fixes that.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>